beam search代码移至utils
